+++ math/src/main/java/coppercore/math/InterpolateDouble.java
